feat: 채팅발생시 알림 호출 메서드 추가 #93
- 채팅 전송에 문제가 발생하지 않도록 try/catch로 알림 호출 분리
- 알림 테스트용 컨트롤러 이름 변경
